backend/requirements.txt
# JAC Learning Platform - Production Dependencies
# Real-time WebSocket support
channels==4.3.2
channels-redis==4.1.0

aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.11.0
ast-decompiler==0.8.0
attrs==25.4.0
audioread==3.1.0
beautifulsoup4==4.14.2
billiard==4.2.3
brotli==1.2.0 ; platform_python_implementation == 'CPython'
brotlicffi==1.1.0.0 ; platform_python_implementation != 'CPython'
celery==5.5.3
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6 ; sys_platform == 'win32'
contourpy==1.3.3
cryptography==46.0.3
cssselect2==0.8.0
cycler==0.12.1
decorator==5.2.1
django==5.2.8
django-allauth==65.13.1
django-cors-headers==4.9.0
django-debug-toolbar==6.1.0
django-extensions==4.1
django-js-asset==3.1.2
django-mptt==0.18.0
django-redis==5.2.0
django-celery-beat==2.8.1
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-filter==25.2
google-generativeai==0.8.3
docstring-parser==0.17.0
docx2txt==0.9
drf-spectacular==0.29.0
edge-tts==7.2.3
et-xmlfile==2.0.0
factory-boy==3.3.3
faker==38.2.0
fonttools==4.60.1
frozenlist==1.8.0
greenlet==3.2.4
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
idna==3.11
inflection==0.5.1
iniconfig==2.3.0
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
kombu==5.5.4
lazy-loader==0.4
librosa==0.11.0
llvmlite==0.45.1
lxml==6.0.2
matplotlib==3.10.7
msgpack==1.1.2
multidict==6.7.0
mypy==1.18.2
mypy-extensions==1.1.0
networkx==3.5
nltk==3.9.2
numba==0.62.1
numpy==2.2.6
opencv-python==4.12.0.88
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pathspec==0.12.1
patsy==1.0.2
pdfminer-six==20251107
pdfplumber==0.11.8
pillow==12.0.0
platformdirs==4.5.0
playwright==1.56.0
pluggy==1.6.0
pooch==1.8.2
prompt-toolkit==3.0.52
propcache==0.4.1
psycopg2-binary==2.9.11
pycparser==2.23 ; implementation_name != 'PyPy'
pydantic==2.12.4
pydantic-core==2.41.5
pydyf==0.11.0
pyee==13.0.0
pygments==2.19.2
pyjwt==2.10.1
pymupdf==1.26.6
pyparsing==3.2.5
pypdf2==3.0.1
pypdfium2==5.0.0
pyphen==0.17.2
pytest==9.0.1
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.2.0
python-pptx==1.0.2
pytz==2025.2
pyyaml==6.0.3
redis==5.2.1
referencing==0.37.0
regex==2025.11.3
reportlab==4.4.5
requests==2.32.5
rpds-py==0.29.0
scikit-learn==1.7.2
scipy==1.16.3
seaborn==0.13.2
sentry-sdk==2.45.0
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.8
soxr==1.0.0
sqlparse==0.5.3
statsmodels==0.14.5
tabulate==0.9.0
threadpoolctl==3.6.0
tinycss2==1.5.0
tinyhtml5==2.0.0
tqdm==4.67.1
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
weasyprint==66.0
webencodings==0.5.1
wget==3.2
wordcloud==1.9.4
xlrd==2.0.2
xlsxwriter==3.2.9
yarl==1.22.0
zopfli==0.4.0
